Tests: Moved clipboard and enter integration tests of DocumentListEditing to another file.
